Atualizações gerais

+++ app/Http/Controllers/AvaliacaoController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers;
 use App\Mail\SendSolicitacaoStatus;
 use App\Models\AvaliacaoIndividual;
 use App\Models\Avaliacao;
+use App\Models\HistoricoSolicitacao;
 use App\Models\Solicitacao;
 use App\Models\User;
 use App\Models\Licenca;
@@ -18,107 +19,125 @@ class AvaliacaoController extends Controller
     public function aprovarSolicitacao(Request $request)
     {
         $solicitacao = Solicitacao::find($request->solicitacao_id);
-        $avaliador = User::find(Auth::user()->id);
         $avaliacao = Avaliacao::find($request->avaliacao_id);
+        if (Auth::user()->id == 2) {
 
-        if (!$solicitacao->avaliacao_individual || !$solicitacao->avaliacao_individual->status) {
-            return redirect()->back()->with('fail', 'Avaliação Pendente');
+            $avaliador = User::find(Auth::user()->id);
+
+            if (!$solicitacao->avaliacao_individual || !$solicitacao->avaliacao_individual->status) {
+                return redirect()->back()->with('fail', 'Avaliação Pendente');
+            }
+
+            if (!$solicitacao->responsavel->avaliacao_individual || !$solicitacao->responsavel->avaliacao_individual->status) {
+                return redirect()->back()->with('fail', 'Avaliação Pendente');
+            }
+
+
+            $colaboradorAvaliacao = $solicitacao->avaliacao->first()->avaliacao_individual->where('tipo', 2)->first();
+            if (!$colaboradorAvaliacao || !$colaboradorAvaliacao->status) {
+                return redirect()->back()->with('fail', 'Avaliação Pendente');
+            }
+
+
+            if ($solicitacao->dadosComplementares) {
+                $avaliacaoDadosComplementares = $solicitacao->dadosComplementares->avaliacao_individual;
+                if (!$avaliacaoDadosComplementares || !$avaliacaoDadosComplementares->status) {
+                    return redirect()->back()->with('fail', 'Avaliação Pendente');
+                }
+            }
+
+            foreach ($solicitacao->modelosAnimais as $modeloAnimal) {
+
+                $avaliacaoModelo = AvaliacaoIndividual::where('avaliacao_id', $avaliacao->id)
+                    ->where('modelo_animal_id', $modeloAnimal->id)
+                    ->first();
+
+                if (!$avaliacaoModelo || $avaliacaoModelo->status === null) {
+                    return redirect()->back()->with('fail', 'Avaliação Pendente');
+                }
+
+                $avaliacaoPlanejamento = AvaliacaoIndividual::where('avaliacao_id', $avaliacao->id)
+                    ->where('planejamento_id', $modeloAnimal->planejamento->id)
+                    ->first();
+
+                if (!$avaliacaoPlanejamento || $avaliacaoPlanejamento->status === null) {
+                    return redirect()->back()->with('fail', 'Avaliação Pendente');
+                }
+
+                $avaliacaoCondicoesAnimal = AvaliacaoIndividual::where('avaliacao_id', $avaliacao->id)
+                    ->where('condicoes_animal_id', $modeloAnimal->planejamento->condicoesAnimal->id)
+                    ->first();
+
+                if (!$avaliacaoCondicoesAnimal || $avaliacaoCondicoesAnimal->status === null) {
+                    return redirect()->back()->with('fail', 'Avaliação Pendente');
+                }
+
+                $avaliacaoProcedimento = AvaliacaoIndividual::where('avaliacao_id', $avaliacao->id)
+                    ->where('procedimento_id', $modeloAnimal->planejamento->procedimento->id)
+                    ->first();
+
+                if (!$avaliacaoProcedimento || $avaliacaoProcedimento->status === null) {
+                    return redirect()->back()->with('fail', 'Avaliação Pendente');
+                }
+
+                $avaliacaoOperacao = AvaliacaoIndividual::where('avaliacao_id', $avaliacao->id)
+                    ->where('operacao_id', $modeloAnimal->planejamento->operacao->id)
+                    ->first();
+
+                if (!$avaliacaoOperacao || $avaliacaoOperacao->status === null) {
+                    return redirect()->back()->with('fail', 'Avaliação Pendente');
+                }
+
+                $avaliacaoEutanasia = AvaliacaoIndividual::where('avaliacao_id', $avaliacao->id)
+                    ->where('eutanasia_id', $modeloAnimal->planejamento->eutanasia->id)
+                    ->first();
+
+                if (!$avaliacaoEutanasia || $avaliacaoEutanasia->status === null) {
+                    return redirect()->back()->with('fail', 'Avaliação Pendente');
+                }
+
+                $avaliacaoResultado = AvaliacaoIndividual::where('avaliacao_id', $avaliacao->id)
+                    ->where('resultado_id', $modeloAnimal->planejamento->resultado->id)
+                    ->first();
+
+                if (!$avaliacaoResultado || $avaliacaoResultado->status === null) {
+                    return redirect()->back()->with('fail', 'Avaliação Pendente');
+                }
+            }
+
+            Avaliacao::where('solicitacao_id', $solicitacao->id)->where('user_id', '!=', $avaliador->id)->delete();
+            $avaliacao->status = 'aprovado_avaliador';
+            $solicitacao->status = 'avaliado';
+            $solicitacao->update();
+            $avaliacao->update();
         }
 
-        if (!$solicitacao->responsavel->avaliacao_individual || !$solicitacao->responsavel->avaliacao_individual->status) {
-            return redirect()->back()->with('fail', 'Avaliação Pendente');
+        if (Auth::user()->id == 1) {
+            $avaliacao->status = 'aprovado_colegiado';
+            $avaliacao->update();
+            //Criação da licença
+            $licenca = new Licenca();
+            $licenca->inicio = $request->inicio;
+            $licenca->fim = $request->fim;
+            $licenca->codigo = "CEUAUFAPE" . date('Y', strtotime($solicitacao->updated_at)) . date('m', strtotime($solicitacao->updated_at)) . date('d', strtotime($solicitacao->updated_at)) . $solicitacao->id;
+            // $licenca->codigo = strtoupper(hash('ripemd160', $solicitacao->id.$request->inicio.$request->fim));
+            $licenca->avaliacao_id = $avaliacao->id;
+            $licenca->save();
         }
 
+        $historico = new HistoricoSolicitacao();
+        $historico->solicitacao_id = $request->solicitacao_id;
+        $historico->status_solicitacao = $avaliacao->status;
+        $historico->nome_usuario_modificador = Auth::user()->name;
 
-        $colaboradorAvaliacao = $solicitacao->avaliacao->first()->avaliacao_individual->where('tipo', 2)->first();
-        if (!$colaboradorAvaliacao || !$colaboradorAvaliacao->status) {
-            return redirect()->back()->with('fail', 'Avaliação Pendente');
-        }
-
-
-        if ($solicitacao->dadosComplementares) {
-            $avaliacaoDadosComplementares = $solicitacao->dadosComplementares->avaliacao_individual;
-            if (!$avaliacaoDadosComplementares || !$avaliacaoDadosComplementares->status) {
-                return redirect()->back()->with('fail', 'Avaliação Pendente');
-            }
-        }
-
-        foreach ($solicitacao->modelosAnimais as $modeloAnimal) {
-
-            $avaliacaoModelo = AvaliacaoIndividual::where('avaliacao_id', $avaliacao->id)
-                ->where('modelo_animal_id', $modeloAnimal->id)
-                ->first();
-
-            if (!$avaliacaoModelo || $avaliacaoModelo->status === null) {
-                return redirect()->back()->with('fail', 'Avaliação Pendente');
-            }
-
-            $avaliacaoPlanejamento = AvaliacaoIndividual::where('avaliacao_id', $avaliacao->id)
-                ->where('planejamento_id', $modeloAnimal->planejamento->id)
-                ->first();
-
-            if (!$avaliacaoPlanejamento || $avaliacaoPlanejamento->status === null) {
-                return redirect()->back()->with('fail', 'Avaliação Pendente');
-            }
-
-            $avaliacaoCondicoesAnimal = AvaliacaoIndividual::where('avaliacao_id', $avaliacao->id)
-                ->where('condicoes_animal_id', $modeloAnimal->planejamento->condicoesAnimal->id)
-                ->first();
-
-            if (!$avaliacaoCondicoesAnimal || $avaliacaoCondicoesAnimal->status === null) {
-                return redirect()->back()->with('fail', 'Avaliação Pendente');
-            }
-
-            $avaliacaoProcedimento = AvaliacaoIndividual::where('avaliacao_id', $avaliacao->id)
-                ->where('procedimento_id', $modeloAnimal->planejamento->procedimento->id)
-                ->first();
-
-            if (!$avaliacaoProcedimento || $avaliacaoProcedimento->status === null) {
-                return redirect()->back()->with('fail', 'Avaliação Pendente');
-            }
-
-            $avaliacaoOperacao = AvaliacaoIndividual::where('avaliacao_id', $avaliacao->id)
-                ->where('operacao_id', $modeloAnimal->planejamento->operacao->id)
-                ->first();
-
-            if (!$avaliacaoOperacao || $avaliacaoOperacao->status === null) {
-                return redirect()->back()->with('fail', 'Avaliação Pendente');
-            }
-
-            $avaliacaoEutanasia = AvaliacaoIndividual::where('avaliacao_id', $avaliacao->id)
-                ->where('eutanasia_id', $modeloAnimal->planejamento->eutanasia->id)
-                ->first();
-
-            if (!$avaliacaoEutanasia || $avaliacaoEutanasia->status === null) {
-                return redirect()->back()->with('fail', 'Avaliação Pendente');
-            }
-
-            $avaliacaoResultado = AvaliacaoIndividual::where('avaliacao_id', $avaliacao->id)
-                ->where('resultado_id', $modeloAnimal->planejamento->resultado->id)
-                ->first();
-
-            if (!$avaliacaoResultado || $avaliacaoResultado->status === null) {
-                return redirect()->back()->with('fail', 'Avaliação Pendente');
-            }
-        }
-
-        Avaliacao::where('solicitacao_id', $solicitacao->id)->where('user_id', '!=', $avaliador->id)->delete();
-        $avaliacao->status = 'aprovado';
-        $solicitacao->status = 'avaliado';
-        $solicitacao->update();
-        $avaliacao->update();
-        //Criação da licença
-        $licenca = new Licenca();
-        $licenca->inicio = $request->inicio;
-        $licenca->fim = $request->fim;
-        $licenca->codigo = "CEUAUFAPE" . date('Y', strtotime($solicitacao->updated_at)) . date('m', strtotime($solicitacao->updated_at)) . date('d', strtotime($solicitacao->updated_at)) . $solicitacao->id;
-        // $licenca->codigo = strtoupper(hash('ripemd160', $solicitacao->id.$request->inicio.$request->fim));
-        $licenca->avaliacao_id = $avaliacao->id;
-        $licenca->save();
+        $historico->save();
 
         Mail::to($solicitacao->responsavel->contato->email)->send(new SendSolicitacaoStatus($solicitacao->responsavel, $avaliacao));
 
-        return redirect(route('solicitacao.avaliador.index'));
+        if(Auth::user()->id == 1){
+            return redirect(route('solicitacao.admin.index'));
+        }
+            return redirect(route('solicitacao.avaliador.index'));
     }
 
     public function aprovarPendenciaSolicitacao(Request $request)
@@ -215,6 +234,13 @@ class AvaliacaoController extends Controller
         $solicitacao->update();
         $avaliacao->update();
 
+        $historico = new HistoricoSolicitacao();
+        $historico->solicitacao_id = $request->solicitacao_id;
+        $historico->status_solicitacao = $avaliacao->status;
+        $historico->nome_usuario_modificador = Auth::user()->name;
+
+        $historico->save();
+
         Mail::to($solicitacao->responsavel->contato->email)->send(new SendSolicitacaoStatus($solicitacao->responsavel, $avaliacao));
 
         return redirect(route('solicitacao.avaliador.index'));
@@ -308,14 +334,26 @@ class AvaliacaoController extends Controller
             }
         }
 
-        Avaliacao::where('solicitacao_id', $solicitacao->id)->where('user_id', '!=', $avaliador->id)->delete();
+        if (Auth::user()->id == 2){
+            Avaliacao::where('solicitacao_id', $solicitacao->id)->where('user_id', '!=', $avaliador->id)->delete();
+        }
         $avaliacao->status = 'reprovado';
         $solicitacao->status = 'avaliado';
         $solicitacao->update();
         $avaliacao->update();
 
+        $historico = new HistoricoSolicitacao();
+        $historico->solicitacao_id = $request->solicitacao_id;
+        $historico->status_solicitacao = $avaliacao->status;
+        $historico->nome_usuario_modificador = Auth::user()->name;
+
+        $historico->save();
+
         Mail::to($solicitacao->responsavel->contato->email)->send(new SendSolicitacaoStatus($solicitacao->responsavel, $avaliacao));
 
+        if(Auth::user()->id == 1){
+            return redirect(route('solicitacao.admin.index'));
+        }
         return redirect(route('solicitacao.avaliador.index'));
     }
 }
